Enable CodeMirror search.
Ctrl-F to search, but it is not interactive, need to press enter. Ctrl-G
for search next. Alt-G to jump to line.

This was needed, as we set CodeMirror to render the code text on-demand
only for speedy loading, but that breaks native browser search (the
searched text might not be rendered, even if it is contained in the file
somewhere).