Fix layout issue + many small things.
The codemirror view didn't properly expand to the desired height.
Changed how to calculate that now, and on what event to update it.

FileTree highlight now works from both nav and ref click.

Axios was buggy, if json source was loaded, even if header specified
text/plain, it tried to parse to json, breaking the source view. Worked
around.

Copy-pasta ref display for declarations, so if zoekt-underhood sends
them, we can show.